requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
proxmoxer==2.2.0
psutil==7.1.0
nest_asyncio==1.6.0
